requirements.txt
langchain==0.3.21
langchain-community==0.3.20
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-chroma==0.1.4
langgraph==0.3.21
langgraph-supervisor==0.0.15
chromadb==0.5.23
fastapi==0.115.11
uvicorn==0.34.0
streamlit==1.43.2
python-dotenv==1.0.1
pandas==2.2.3
sqlalchemy==2.0.39
langsmith==0.3.15
sentence-transformers==3.4.1
